improvements of refactored site
On this version, the stylesheet hints.css was created and it would be used to all hints pages.
The page hintsHtml5Css3.html was created and was inserted on hint
Some more pages were created as provision on _links folder but they were not finished yet.
The pages mainPython.html and mainHtml5Css3.html were changed a little (only details)